UBUNTU: azure: force rdma_ucm to be loaded when hv_network_direct_* is loaded
BugLink: http://bugs.launchpad.net/bugs/1712915

Create a soft dependency to automatically load rdma_ucm when a Hyper-V
RDMA driver is loaded. Since the RDMA drivers are usually loaded by
WALinuxAgent, the agent must be installed on the system for rdma_ucm to
be loaded during the boot.

The wildcard used in the softdep directive covers all RDMA driver
versions and it's also triggered when the driver is loaded via an alias,
i.e. "hv_network_direct".

Signed-off-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>