UBUNTU: [Config] Add annotation for CONFIG_CONTEXT_TRACKING_FORCE
This option is not selectable in the master kernel, but it is in
linux-azure because it has CONFIG_VIRT_CPU_ACCOUNTING_GEN=y. I
suspect this is not the value we want for that option, but as
this is just being noticed when working on the 20.04 freeze
kernel I'm going to leave it unchanged and flag
CONFIG_CONTEXT_TRACKING_FORCE for review.

Ignore: yes
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>