UBUNTU: SAUCE: Drivers: hv: vmbus: Restrict vmbus_devices on isolated guests
BugLink: http://bugs.launchpad.net/bugs/1915146

Only the VSCs or ICs that have been hardened and that are critical for
the successful adoption of Confidential VMs should be allowed if the
guest is running isolated.  This change reduces the footprint of the
code that will be exercised by Confidential VMs and hence the exposure
to bugs and vulnerabilities.

Signed-off-by: Andrea Parri (Microsoft) <parri.andrea@gmail.com>
Reviewed-by: Michael Kelley <mikelley@microsoft.com>
Link: https://lore.kernel.org/r/20210201144814.2701-3-parri.andrea@gmail.com
Signed-off-by: Wei Liu <wei.liu@kernel.org>
(cherry picked from commit a2528a10a5bd39306f47c4db33527df215e4a22b
 from https://git.kernel.org/pub/scm/linux/kernel/git/hyperv/linux.git)
Signed-off-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>