UBUNTU: [Config] do not enforce DMESG_RESTRICT
Andrea Righi writes for focal:linux-hwe-5.8:
DMESG_RESTRICT is enforced in groovy onward, but it would cause problems
if enabled in focal, because it's missing the necessary user-space
package updates, so make sure this option is not enabled.

Unlike the commit in focal:linux-hwe-5.8, here annotations have to be
updated as well.

Signed-off-by: Krzysztof Kozlowski <krzysztof.kozlowski@canonical.com>