UBUNTU: [Packaging] Update the auto generated control file fields
Conflict only with the signed counterpart of this package since
there's no need to conflict with the image packages from the base
kernel.

Drop the Recommends clause, since that is only used to bring grub
packages.

And finally drop any mention to initramfs (mainly from Suggests) since
the initrd is already included to the EFI binary.

Signed-off-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>