UBUNTU: azure: [Config] disable KFENCE
BugLink: https://bugs.launchpad.net/bugs/1946001

Disable KFENCE as a workaround to prevent potential kernel panics that
seem to happen in nested KVM environments (e.g., systemd autopackage
test).

Disabling this feature also allows to remove the overhead of the
periodic sampling performed by KFENCE. The regression potential is that
we may receive memory corruption bug reports (that were previously
detected explicitly by KFENCE), but in such case we have always the
option to provide a test kernel with KASAN enabled, that provides a
better coverage for this kind of bugs.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>