UBUNTU: [Config] Restore BTF debug info in modules
(Based on base kernel commit 7300f50c623d19596a7d8f45d664e596c028d03a)

One has to be careful to have the right PPA enabled in the cranky chroot
session which updates the config. Because pahole (which comes with
dwarves) has to be at a version which is only in updates but not in
security.

Ignore: yes
Signed-off-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>